+++ BirbCore/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.1.0
 Updated for Stardew Valley 1.6.13
+### Added
+- TriggerAction delegate
 ### Fixed
 - Split-screen issue with StatChanged Event annotation